refactored commands/data/methods
