+++ data/config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    prefix: "someone!",
+    creatorid: '492079026089885708' 
+}
