+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "Someone",
-  "version": "1.2.2",
+  "version": "2.0.1",
   "description": "Someone bot for Discord. Brings back Discord's 2018 April Fools joke.",
   "main": "index.js",
   "scripts": {
@@ -12,7 +12,7 @@
     "url": "git+https://github.com/ApocalypseCalculator/Someone"
   },
   "author": "ApocalypseCalculator",
-  "license": "MIT",
+  "license": "CC-BY-NC-SA-4.0",
   "bugs": {
     "url": "https://github.com/ApocalypseCalculator/Someone/issues"
   },
