--- token.js
@@ -1,1 +1,0 @@
-module.exports = "put your token here!";
